+++ datalogger_jupyterlab/__init__.py
@@ -1,5 +1,0 @@
-"""JupyterLab extension configuration."""
-
-
-def _jupyter_labextension_paths():
-    return [{"src": "../labextension", "dest": "datalogger-jupyterlab"}]