+++ src/constants.ts
@@ -1,9 +1,5 @@
+export const PACKAGE_NAME = "jupyterlab-datalogger";
+
 // Mimetypes
-export const PDF_MIMETYPE = "application/pdf";
 export const JSON_MIMETYPE = "application/json";
 export const NETCDF_MIMETYPE = "application/x-netcdf";
-
-// PDF preview parameters
-export const PDF_PREVIEW_WIDTH = 250;
-export const PDF_PREVIEW_BORDER_WIDTH = 1;
-export const PDF_PREVIEW_GAP = 10;