+++ src/index.ts
@@ -1,16 +1,3 @@
-import applyModifications from "@/modifications";
-import {
-  netcdfFileTypePlugin,
-  dataloggerLoadCodePlugin,
-  pdfPreviewPlugin,
-  openWarningPlugin,
-} from "@/plugins";
+import { netcdfFileTypePlugin, dataloggerLoadCodePlugin } from "@/plugins";
 
-applyModifications();
-
-export default [
-  netcdfFileTypePlugin,
-  dataloggerLoadCodePlugin,
-  pdfPreviewPlugin,
-  openWarningPlugin,
-];
+export default [netcdfFileTypePlugin, dataloggerLoadCodePlugin];