+++ src/plugins/index.ts
@@ -1,4 +1,2 @@
-export { default as netcdfFileTypePlugin } from "./netcdfFileTypePlugin";
-export { default as dataloggerLoadCodePlugin } from "./dataloggerLoadCodePlugin";
-export { default as pdfPreviewPlugin } from "./pdfPreviewPlugin";
-export { default as openWarningPlugin } from "./openWarningPlugin";
+export * from "./netcdfFileTypePlugin";
+export * from "./dataloggerLoadCodePlugin";