+++ src/utils/index.ts
@@ -1,3 +1,2 @@
-export { default as generateLoadCode } from "./generateLoadCode";
-export { default as addToNotebook } from "./addToNotebook";
-export { default as calcPdfPreviewPosition } from "./calcPdfPreviewPosition";
+export * from "./generateLoadCode";
+export * from "./addToNotebook";